# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pop The Question API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: daily
    description: Daily games
  - name: rooms
    description: Hosted game rooms
  - name: questions
    description: Question packs

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /daily/three-strikes:
    get:
      operationId: getTodayThreeStrikes
      tags: [daily]
      summary: Get today's Three Strikes challenge
      responses:
        "200":
          description: Today's challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeStrikesChallenge"
        "404":
          description: No challenge for today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /daily/three-strikes/archive:
    get:
      operationId: getThreeStrikesArchive
      tags: [daily]
      summary: Get archive of Three Strikes challenges
      responses:
        "200":
          description: List of past challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThreeStrikesSummary"

  /daily/three-strikes/{id}:
    get:
      operationId: getThreeStrikesById
      tags: [daily]
      summary: Get a specific Three Strikes challenge by ID (for archive replay)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The challenge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThreeStrikesChallenge"
        "404":
          description: Challenge not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /daily/crossword:
    get:
      operationId: getTodayCrossword
      tags: [daily]
      summary: Get today's crossword puzzle
      responses:
        "200":
          description: Today's crossword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrosswordPuzzle"
        "404":
          description: No puzzle for today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /daily/crossword/archive:
    get:
      operationId: getCrosswordArchive
      tags: [daily]
      summary: Get archive of crossword puzzles
      responses:
        "200":
          description: List of past puzzles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CrosswordSummary"

  /daily/pop-box:
    get:
      operationId: getTodayPopBox
      tags: [daily]
      summary: Get today's Pop Box grid
      responses:
        "200":
          description: Today's Pop Box
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopBoxGrid"
        "404":
          description: No Pop Box for today
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /daily/pop-box/archive:
    get:
      operationId: getPopBoxArchive
      tags: [daily]
      summary: Get archive of Pop Box grids
      responses:
        "200":
          description: List of past Pop Box grids
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PopBoxSummary"

  /daily/pop-box/{id}:
    get:
      operationId: getPopBoxById
      tags: [daily]
      summary: Get a specific Pop Box grid by ID (for archive replay)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The grid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopBoxGrid"
        "404":
          description: Grid not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /daily/pop-box/{id}/answers:
    get:
      operationId: getPopBoxAnswers
      tags: [daily]
      summary: Get all valid celebrities for each cell of a Pop Box grid, plus rarity stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cell answers and rarity stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopBoxAnswers"
        "404":
          description: Grid not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /daily/pop-box/{id}/guess:
    post:
      operationId: popBoxGuess
      tags: [daily]
      summary: Submit a celebrity guess for a Pop Box cell
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PopBoxGuessRequest"
      responses:
        "200":
          description: Guess result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopBoxGuessResult"
        "404":
          description: Grid not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /daily/status:
    get:
      operationId: getDailyStatus
      tags: [daily]
      summary: Get today's completion status summary (wow endpoint)
      responses:
        "200":
          description: Today's game status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyStatus"

  /rooms/create:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new hosted game room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomRequest"
      responses:
        "200":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomCreated"

  /rooms/{code}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room info
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /questions/pop-the-question:
    get:
      operationId: getPopTheQuestionPrompts
      tags: [questions]
      summary: Get Pop the Question prompts
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuestionPrompt"

  /questions/roast-roulette:
    get:
      operationId: getRoastRouletteQuestions
      tags: [questions]
      summary: Get Roast Roulette questions
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoastQuestion"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    ThreeStrikesAnswer:
      type: object
      properties:
        display:
          type: string
        hint:
          type: string
        correct:
          type: array
          items:
            type: string
      required:
        - display
        - hint
        - correct

    ThreeStrikesChallenge:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        title:
          type: string
        prompt:
          type: string
        totalCount:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ThreeStrikesAnswer"
      required:
        - id
        - date
        - title
        - prompt
        - totalCount
        - answers

    ThreeStrikesSummary:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        title:
          type: string
        prompt:
          type: string
        totalCount:
          type: integer
      required:
        - id
        - date
        - title
        - prompt
        - totalCount

    CrosswordPuzzle:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        grid:
          type: array
          items:
            type: array
            items:
              type: string
        blackSquares:
          type: array
          items:
            type: array
            items:
              type: integer
        cluesAcross:
          type: object
          additionalProperties:
            type: string
        cluesDown:
          type: object
          additionalProperties:
            type: string
      required:
        - id
        - date
        - grid
        - blackSquares
        - cluesAcross
        - cluesDown

    CrosswordSummary:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
      required:
        - id
        - date

    DailyStatus:
      type: object
      properties:
        date:
          type: string
        threeStrikesAvailable:
          type: boolean
        crosswordAvailable:
          type: boolean
        popBoxAvailable:
          type: boolean
        threeStrikesTitle:
          type: string
        crosswordDate:
          type: string
        popBoxDate:
          type: string
      required:
        - date
        - threeStrikesAvailable
        - crosswordAvailable
        - popBoxAvailable

    PopBoxCategory:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        group:
          type: string
      required:
        - id
        - label
        - group

    PopBoxGrid:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        rowCategories:
          type: array
          items:
            $ref: "#/components/schemas/PopBoxCategory"
        columnCategories:
          type: array
          items:
            $ref: "#/components/schemas/PopBoxCategory"
      required:
        - id
        - date
        - difficulty
        - rowCategories
        - columnCategories

    PopBoxSummary:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        difficulty:
          type: string
      required:
        - id
        - date
        - difficulty

    PopBoxValidCelebrity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        guessCount:
          type: integer
        rarityPercent:
          type: number
      required:
        - id
        - name
        - guessCount
        - rarityPercent

    PopBoxCellAnswers:
      type: object
      properties:
        squareIndex:
          type: integer
        rowCategoryId:
          type: string
        columnCategoryId:
          type: string
        validCelebrities:
          type: array
          items:
            $ref: "#/components/schemas/PopBoxValidCelebrity"
      required:
        - squareIndex
        - rowCategoryId
        - columnCategoryId
        - validCelebrities

    PopBoxAnswers:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        totalGuesses:
          type: integer
        cells:
          type: array
          items:
            $ref: "#/components/schemas/PopBoxCellAnswers"
      required:
        - id
        - date
        - totalGuesses
        - cells

    PopBoxGuessRequest:
      type: object
      properties:
        squareIndex:
          type: integer
          minimum: 0
          maximum: 8
        guess:
          type: string
          minLength: 1
          maxLength: 100
      required:
        - squareIndex
        - guess

    PopBoxGuessResult:
      type: object
      properties:
        correct:
          type: boolean
        reason:
          type: string
          nullable: true
          enum: [unknown_celebrity, wrong_cell, null]
        celebrityId:
          type: string
          nullable: true
        celebrityName:
          type: string
          nullable: true
        rarityPercent:
          type: number
          nullable: true
      required:
        - correct

    CreateRoomRequest:
      type: object
      properties:
        gameType:
          type: string
          enum: [pop-the-question, roast-roulette, pub-quiz, jeopardy, wheel-of-fortune]
        demo:
          type: boolean
        packId:
          type: string
      required:
        - gameType

    RoomCreated:
      type: object
      properties:
        roomCode:
          type: string
        gameType:
          type: string
      required:
        - roomCode
        - gameType

    Room:
      type: object
      properties:
        code:
          type: string
        gameType:
          type: string
        status:
          type: string
          enum: [lobby, playing, finished]
        playerCount:
          type: integer
      required:
        - code
        - gameType
        - status
        - playerCount

    QuestionPrompt:
      type: object
      properties:
        id:
          type: integer
        prompt:
          type: string
        category:
          type: string
      required:
        - id
        - prompt

    RoastQuestion:
      type: object
      properties:
        id:
          type: integer
        question:
          type: string
        color:
          type: string
      required:
        - id
        - question