# lib/api-zod/src/generated/types/dailyStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pop The Question API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DailyStatus {
  date: string;
  threeStrikesAvailable: boolean;
  crosswordAvailable: boolean;
  popBoxAvailable: boolean;
  threeStrikesTitle?: string;
  crosswordDate?: string;
  popBoxDate?: string;
}