# lib/api-zod/src/generated/types/popBoxAnswers.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pop The Question API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PopBoxCellAnswers } from "./popBoxCellAnswers";

export interface PopBoxAnswers {
  id: string;
  date: string;
  totalGuesses: number;
  cells: PopBoxCellAnswers[];
}