# lib/api-zod/src/generated/types/popBoxCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pop The Question API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PopBoxCategory {
  id: string;
  label: string;
  group: string;
}