# lib/api-zod/src/generated/types/popBoxCellAnswers.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pop The Question API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PopBoxValidCelebrity } from "./popBoxValidCelebrity";

export interface PopBoxCellAnswers {
  squareIndex: number;
  rowCategoryId: string;
  columnCategoryId: string;
  validCelebrities: PopBoxValidCelebrity[];
}