# lib/api-zod/src/generated/types/popBoxGuessRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pop The Question API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PopBoxGuessRequest {
  /**
   * @minimum 0
   * @maximum 8
   */
  squareIndex: number;
  /**
   * @minLength 1
   * @maxLength 100
   */
  guess: string;
}