# lib/api-zod/src/generated/types/popBoxGuessResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pop The Question API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PopBoxGuessResultReason } from "./popBoxGuessResultReason";

export interface PopBoxGuessResult {
  correct: boolean;
  reason?: PopBoxGuessResultReason;
  celebrityId?: string | null;
  celebrityName?: string | null;
  rarityPercent?: number | null;
}