# lib/api-zod/src/generated/types/popBoxSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pop The Question API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PopBoxSummary {
  id: string;
  date: string;
  difficulty: string;
}