# lib/api-zod/src/generated/types/popBoxValidCelebrity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pop The Question API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PopBoxValidCelebrity {
  id: string;
  name: string;
  guessCount: number;
  rarityPercent: number;
}